requirements.txt
fastapi
uvicorn
google-generativeai>=0.7.0
supabase
pydantic